requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
python-dotenv
passlib
python-jose[cryptography]
email-validator
jinja2
paho-mqtt
requests
